Simulation testing no longer uses parameterised tests as this was inelegant and there were difficulties getting the simulations to iterate once per run of all of the tests (without access to `@BeforeParams` introduced in junit 4.13). Instead we use a new class inheriting from Suite that creates a set of tests to be run, one for each iteration, and manages the iteration of the simulations.
